Moved BAV test data file into a new data folder
